fix(ISSUE2): standardization of character route
Change the name of the file from 'Character.cs' to 'CharacterRoute.cs'.
Add try catch expression.
